fix attribute lists for LDAP queries
LDAP does not know of an attribute named 'dn'.
To get only the DN in a search, the attriibute list to use is '1.1'.
On all other cases, the DN of the entries found is automaticlly
part of the result set too.

FossilOrigin-Name: 7fd8da252b548c1dfe609c0b87f68370f316a53cdb88954e24353587169181bf