Allow default attributes to be set via config and cmdline
This commit changes the default attribute list from a fixed '*'
to a configurable list, defaulting to ['*'].

It allows one to use config file or command line to set default
attributes to retrieve and display - i.e. to set default attributes
to '+' or to a list of multiple specific attributes.

Printing or changing this value through env/setenv is currently not
supported because env/setenv only work properly with string values
while attributes is a list.